requirements.txt
fastapi>=0.104.1
openai==0.28.1
python-dotenv>=1.0.0
backoff>=2.2.1
requests>=2.31.0
starlette>=0.40.0,<0.46.0
jinja2>=3.0.0  # For Jinja2Templates
pydantic>=2.5.2
uvicorn>=0.23.1
itsdangerous>=2.1.2
python-multipart>=0.0.6
IPython>=8.13.2
